% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures-of-assortativity.R
\name{calculate_ev_ratio}
\alias{calculate_ev_ratio}
\title{Calculate q statistic}
\usage{
calculate_ev_ratio(mat)
}
\arguments{
\item{mat}{An n x n transmission matrix}
}
\value{
The value of q
}
\description{
Calculates the q from \insertCite{KeelingRohani2008-10.2307/j.ctvcm4gk0.6;textual}{assortativity}
}
\examples{
mat <- matrix(data = c(0.25, 0.1, 0.15,
0.05, 0.1, 0.07,
0.08, 0.09, 0.11),
ncol = 3, nrow = 3, byrow = TRUE)
calculate_ev_ratio(mat)
# 3.994067
}
\references{
\insertAllCited{}
}
