% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures-of-assortativity.R
\name{calculate_I_stat}
\alias{calculate_I_stat}
\title{Calculate index of disassortativity}
\usage{
calculate_I_stat(pop_frac, mat, classes = NULL)
}
\arguments{
\item{pop_frac}{A vector of length n with the population distribution}

\item{mat}{An n x n transmission matrix}

\item{classes}{A vector of length n with the population sizes}
}
\value{
A list with three variants of the index
\describe{
  \item{I_stat_raw}{The index of disassortativity}
  \item{I_stat_standardised}{The standardised index of disassortativity (`I_stat_raw` standardised to the variance)}
  \item{I_stat_margin}{The index of assortativity standardised with respect to the marginal distribution}
}
}
\description{
Calculates the I-squared from \insertCite{Farrington2009-10.1002/bimj.200800160;textual}{assortativity}
}
\examples{
mat <- matrix(data = c(0.25, 0.1, 0.15,
0.05, 0.1, 0.07,
0.08, 0.09, 0.11),
ncol = 3, nrow = 3, byrow = TRUE)
pop_frac <- c(0.35, 0.56, 0.21)
classes <- c(30, 27, 45)
calculate_I_stat(pop_frac, mat, classes)
# $I_stat_raw
# 40.76486
}
\references{
\insertAllCited{}
}
