% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures-of-assortativity.R
\name{calculate_Q_stat}
\alias{calculate_Q_stat}
\title{Calculate Q statistic}
\usage{
calculate_Q_stat(mat)
}
\arguments{
\item{mat}{An n x n transmission matrix}
}
\value{
The value of Q (takes a value between - 1 / (n - 1) and 1)
}
\description{
Calculates the Q from by \insertCite{Aral1999-10.2105/AJPH.89.6.825;textual}{assortativity}
}
\examples{
mat <- matrix(data = c(0.25, 0.1, 0.15,
0.05, 0.1, 0.07,
0.08, 0.09, 0.11),
ncol = 3, nrow = 3, byrow = TRUE)
calculate_Q_stat(mat)
# -0.27
}
\references{
\insertAllCited{}
}
