% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures-of-assortativity.R
\name{calculate_r_stat}
\alias{calculate_r_stat}
\title{Calculate r statistic}
\usage{
calculate_r_stat(mat)
}
\arguments{
\item{mat}{An n x n transmission matrix}
}
\value{
The value of r (takes a value between 0 and 1)
}
\description{
Calculates the r from \insertCite{Newman2003-10.1103/PhysRevE.67.026126;textual}{assortativity}
}
\examples{
mat <- matrix(data = c(0.25, 0.1, 0.15,
0.05, 0.1, 0.07,
0.08, 0.09, 0.11),
ncol = 3, nrow = 3, byrow = TRUE)
calculate_r_stat(mat)
# 0.1740593
}
\references{
\insertAllCited{}
}
\seealso{
[assortnet::assortment.discrete()]
}
