% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures-of-assortativity.R
\name{plot_mat}
\alias{plot_mat}
\title{Plot the matrix as an annotated heatmap}
\usage{
plot_mat(mat)
}
\description{
Plot the matrix as an annotated heatmap
}
\examples{
mat <- matrix(data = c(0.25, 0.1, 0.15,
0.05, 0.1, 0.07,
0.08, 0.09, 0.11),
ncol = 3, nrow = 3, byrow = TRUE)
plot_mat(mat)
}
\references{
\insertAllCited{}
}
\seealso{
[hhh4contacts::plotC()]
}
