% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures-of-assortativity.R
\name{plot_stat}
\alias{plot_stat}
\title{Plot the statistic on a number line}
\usage{
plot_stat(mat)
}
\description{
Create a number line plot in the style of \insertCite{Garnett1996-10.1097/00007435-199605000-00015;textual}{assortativity}
}
\references{
\insertAllCited{}
}
\seealso{
[calculate_Q_stat()]
#' mat <- matrix(data = c(0.25, 0.1, 0.15,
0.05, 0.1, 0.07,
0.08, 0.09, 0.11),
ncol = 3, nrow = 3, byrow = TRUE)
plot_stat(mat)
}
